develop: add update node method

+++ internal/models/node_record.go
@@ -5,8 +5,8 @@ import "time"
 type NodeRecord struct {
 	NodeId           int64
 	Label            string
-	X                int16
-	Y                int16
+	X                float32
+	Y                float32
 	CreatedTimestamp time.Time
 	UpdatedTimestamp time.Time
 }
