+++ internal/repositories/node_repository.go
@@ -5,7 +5,7 @@ import (
 )
 
 type NodeRepository interface {
-	Save(label string, x int16, y int16) (*models.NodeRecord, error)
+	Save(label string, x float32, y float32) (*models.NodeRecord, error)
 	FindAll() ([]*models.NodeRecord, error)
-	UpdateNode(nodeId int64, label string, x int, y int) error
+	UpdateNode(nodeId int64, label string, x float32, y float32) error
 }
